* src/com/fluendo/jst/Clock.java: (Clock): Time are longs, avoids overflows. * src/com/fluendo/jst/Sink.java: (Sink.Pad): First pass at posting EOS. * src/com/fluendo/player/Cortado.java: (Cortado.handleMessage): Some more debugging about buffering and EOS. * src/com/fluendo/plugin/OggDemux.java: (OggDemux..OggStream.bufferFromPacket): Timestamp outgoing buffers. * src/com/fluendo/plugin/Queue.java: (Queue.updateBuffering), (Queue.Pad): Handle EOS when we are buffering. Fixes #247. * src/com/fluendo/plugin/TheoraDec.java: (TheoraDec..takeHeader), (TheoraDec..getFirstTs), (TheoraDec..Pad): Don't try to convert granulepos if we did not receive headers yet.
git-svn-id: https://code.fluendo.com/flumotion/svn/cortado/trunk@2968 2b0047a9-a6d8-0310-accf-f7200b2a168c

+++ src/com/fluendo/jst/Clock.java
@@ -22,9 +22,9 @@ public abstract class Clock {
   private long adjust;
   private long lastTime;
 
-  public static final int USECOND = 1;
-  public static final int MSECOND = 1000 * USECOND;
-  public static final int SECOND = 1000 * MSECOND;
+  public static final long USECOND = 1;
+  public static final long MSECOND = 1000 * USECOND;
+  public static final long SECOND = 1000 * MSECOND;
 
   /* id types */
   public static final int SINGLE = 0;
